client/requirements.txt
paho-mqtt==1.6.1
requests==2.7.0 # Rest client library

